moved poweredBy into ee/ to make it possible to change it when a whitelabel is applied
